partial cover test

+++ app/pkg/api/v1/accountingapi/request.go
@@ -6,7 +6,7 @@ type NewAccountRequest struct {
 }
 
 type TransactionRequest struct {
-	FromAccountID int `json:"from_account_id"`
-	ToAccountID   int `json:"to_account_id"`
-	Amount        int `json:"amount"`
+	FromAccountID int     `json:"from_account_id"`
+	ToAccountID   int     `json:"to_account_id"`
+	Amount        float64 `json:"amount"`
 }
